Prototyping Functions - Eliminating Name Suffixes
This file is used to build some of the powershell functions that we are
going to need to clean the data. One problem that we run into is having
the suffix be part of the last name field. This function should
eliminate that by deleting whatever is to the right of either a comma
(",") or space (" ").